+++ examples/language_model/llama/export_generation_model.py
@@ -17,7 +17,7 @@ import os
 
 import paddle
 
-from paddlenlp.layers import LoRAConfig, LoRAModel
+from paddlenlp.peft import LoRAConfig, LoRAModel
 from paddlenlp.transformers import AutoModelForCausalLM, AutoTokenizer, LlamaConfig
