+++ paddlenlp/peft/lora/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
+# Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,10 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .prompt_args import *
-from .prompt_model import *
-from .prompt_tokenizer import *
-from .prompt_trainer import *
-from .prompt_utils import *
-from .template import *
-from .verbalizer import *
+from .lora_config import LoRAConfig
+from .lora_layers import (
+    ColumnParallelLoRALinear,
+    ColumnParallelLoRAMergedLinear,
+    LoRALinear,
+    LoRAMergedLinear,
+)
+from .lora_model import LoRAModel
